requirements.txt
fastapi
uvicorn[standard]

# LangChain core
langchain==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28

langchain-google-genai 
google-generativeai

python-dotenv
